# openapi.json
{
  "components": {
...
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "CatDiary REST API v1",
    "title": "猫伴日记 API",
    "version": "1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/families/{id}/invites": {
      "post": {
        "operationId": "FamilyCollaborationController_createInvite",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["family collaboration"]
      }
    },
    "/api/v1/families/{id}/leave": {
      "post": {
        "operationId": "FamilyCollaborationController_leave",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["family collaboration"]
      }
    },
    "/api/v1/families/{id}/members": {
      "get": {
        "operationId": "FamilyCollaborationController_members",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["family collaboration"]
      }
    },
    "/api/v1/families/{id}/members/{memberId}": {
      "delete": {
        "operationId": "FamilyCollaborationController_remove",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "memberId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["family collaboration"]
      },
      "patch": {
        "operationId": "FamilyCollaborationController_changeRole",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "memberId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["family collaboration"]
      }
    },
    "/api/v1/family-invites/{token}/accept": {
      "post": {
        "operationId": "FamilyCollaborationController_accept",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["family collaboration"]
      }
    },
...
    "/api/v1/notification-channels": {
      "get": {
        "operationId": "ChannelsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["notification channels"]
      }
    },
    "/api/v1/notification-channels/feishu": {
      "delete": {
        "operationId": "ChannelsController_remove",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["notification channels"]
      },
      "put": {
        "operationId": "ChannelsController_configure",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["notification channels"]
      }
    },
    "/api/v1/notification-channels/feishu/test": {
      "post": {
        "operationId": "ChannelsController_test",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["notification channels"]
      }
    },
...
    "/api/v1/photos": {
      "get": {
        "operationId": "PhotosController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["photos"]
      },
      "post": {
        "operationId": "PhotosController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["photos"]
      }
    },
    "/api/v1/photos/{id}": {
      "delete": {
        "operationId": "PhotosController_remove",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "if-match",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["photos"]
      },
      "get": {
        "operationId": "PhotosController_get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["photos"]
      },
      "patch": {
        "operationId": "PhotosController_update",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["photos"]
      }
    },
    "/api/v1/photos/{id}/content": {
      "get": {
        "operationId": "PhotosController_content",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["photos"]
      }
    },
    "/api/v1/photos/{id}/set-avatar": {
      "post": {
        "operationId": "PhotosController_setAvatar",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["photos"]
      }
    },
    "/api/v1/photos/{id}/thumbnail": {
      "get": {
        "operationId": "PhotosController_thumbnail",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["photos"]
      }
    },
...
    "/api/v1/uploads/local/{token}": {
      "put": {
        "operationId": "UploadsController_receiveLocal",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["uploads"]
      }
    },
    "/api/v1/uploads/presign": {
      "post": {
        "operationId": "UploadsController_presign",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["uploads"]
      }
    }
  },
  "servers": [],
...
}